Improve code portability (size_t) of debugcallstack and enable more for GCC
In principle the callstack could also be implemented with backtrace, unfortunately we need it also for the leak reporter. This would require to rewrite the leak reporter more fundamentally than it is worth.

Enabled more code for GCC and other compilers and translated the VS 32 bit assembler code to GCC 64 bit assembler code
ctp2_code\ctp\debugtools\debugcallstack.cpp

Since we are dealing with memory addresses they are supposed to be size_t instead of unsigned. On Windows 32 bit unsigned and size_t are the same, but that changes on 64 bit systems, where size_t has to be 64 bit to match the size of pointers.
ctp2_code\ctp\debugtools\debugcallstack.cpp
ctp2_code\ctp\debugtools\debugcallstack.h
ctp2_code\ctp\debugtools\debugmemory.cpp

Added the PVOID type as void*. This type is used by some debug code in debugcallstack.cpp
ctp2_code\os\nowin32\windows.h

+++ ctp2_code/ctp/debugtools/debugcallstack.h
@@ -46,11 +46,11 @@ void DebugCallStack_Close (void);
 void DebugCallStack_Dump (LogClass log_class);
 void DebugCallStack_DumpFrom (LogClass log_class, size_t base_pointer);
 void DebugCallStack_DumpAddress (LogClass log_class, size_t address);
-void DebugCallStack_Save  (unsigned *call_stack, int number, size_t Ebp);
+void DebugCallStack_Save  (size_t *call_stack, int number, size_t Ebp);
 
-void DebugCallStack_Show  (LogClass log_class, unsigned *call_stack, int number);
-void DebugCallStack_ShowToFile  (LogClass log_class, unsigned *call_stack, int number, FILE *file);
-void DebugCallStack_ShowToAltFile  (LogClass log_class, unsigned *call_stack, int number, FILE *file);
+void DebugCallStack_Show  (LogClass log_class, size_t *call_stack, int number);
+void DebugCallStack_ShowToFile  (LogClass log_class, size_t *call_stack, int number, FILE *file);
+void DebugCallStack_ShowToAltFile  (LogClass log_class, size_t *call_stack, int number, FILE *file);
 
 void *Debug_GetFAFirst(void);
 void Debug_SetFAFirst(void *ptr);
@@ -67,12 +67,12 @@ public:
 
 private:
 	int m_maxNumStacks;
-	unsigned m_caller;
+	size_t m_caller;
 	int m_depth;
 	int m_blockSize;
 	int m_numStacks;
-	unsigned *m_stacks;
-	unsigned *m_curStack;
+	size_t *m_stacks;
+	size_t *m_curStack;
 };
 #endif
